Update app.R
Find on Feb 21, 2021 that there is an added space in the county name in the USA FACTS data file that needs to be removed (otherwise filter not working). Also in the USA FACTS file the date format is changed to 2020-01-22 so need to change from mdy (1/22/20) to ymd on line 297.